# 2/first-frontend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/locatii": {
            "get": {
                "tags": [
                    "Locatii"
                ],
                "summary": "Returneaza toate locatiile",
                "description": "Returneaza toate locatiile",
                "operationId": "5f3129f95fc5428b82ec5ee009a3592a",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Locatii"
                ],
                "summary": "Update locatia cu id-ul dat",
                "description": "Update locatia cu id-ul dat",
                "operationId": "d9faba072eeebbc5c15ac5b82a257e8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idLocatie": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "adresa": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Bucuresti, Bd. Iuliu Maniu, Nr. 54"
                                    },
                                    "codPostal": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 63423
                                    },
                                    "email": {
                                        "description": "",
                                        "type": "string",
                                        "example": "bucuresti@first.ro"
                                    },
                                    "nrTelefon": {
                                        "description": "",
                                        "type": "string",
                                        "example": "+40244234340"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Locatii"
                ],
                "summary": "Adauga locatia",
                "description": "Adauga locatia",
                "operationId": "59ab050f56596bd67396d73736b3b0e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "adresa": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Bucuresti, Bd. Iuliu Maniu, Nr. 54"
                                    },
                                    "codPostal": {
                                        "description": "",
                                        "type": "string",
                                        "example": "063423"
                                    },
                                    "email": {
                                        "description": "",
                                        "type": "string",
                                        "example": "bucuresti@first.ro"
                                    },
                                    "nrTelefon": {
                                        "description": "",
                                        "type": "string",
                                        "example": "+40244234340"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/locatii/{id}": {
            "get": {
                "tags": [
                    "Locatii"
                ],
                "summary": "Returneaza locatia cu id-ul dat",
                "description": "Returneaza locatia cu id-ul dat",
                "operationId": "d8a971a21e9eeea46c93a2fdce9bf581",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul locatiaui dorit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locatii"
                ],
                "summary": "Sterge locatia cu id-ul dat",
                "description": "Sterge locatia cu id-ul dat",
                "operationId": "3923373523ac5f47ce43e1f624ee884e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul locatiei dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}